refactor: user profile page (#67)
