+++ src/components/FormChangePassword/FormChangePassword.module.css
@@ -1,0 +1,28 @@
+.modal-password {
+  width: 100%;
+  padding: 1em;
+  background-color: whitesmoke;
+  padding: 1em;
+  border-radius: 5px;
+  border: 1px solid black;
+
+  @media (min-width: 768px) {
+    max-width: 55svw;
+  }
+}
+
+.new-password-fieldset {
+  padding: 1em;
+  margin-top: 0.5em;
+  border-radius: 5px;
+}
+
+.error-msg {
+  color: #ea5555;
+}
+
+.modal-buttons {
+  display: flex;
+  justify-content: space-evenly;
+  margin-top: 1em;
+}
